03/assignment/wafflelargeblog/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==7.1.2
click-didyoumean==0.3.0
cryptography==41.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==3.2
django-allauth==0.58.1
django-braces==1.15.0
django-dotenv==1.4.2
django-rest-auth==0.9.5
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.0
idna==3.4
oauthlib==3.2.2
p==1.4.0
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.7
